+++ lib.commonjs/_tests/utils.js
@@ -1,12 +1,10 @@
 "use strict";
-var __importDefault = (this && this.__importDefault) || function (mod) {
-    return (mod && mod.__esModule) ? mod : { "default": mod };
-};
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.retryIt = exports.stall = exports.log = exports.loadTests = void 0;
-const fs_1 = __importDefault(require("fs"));
-const path_1 = __importDefault(require("path"));
-const zlib_1 = __importDefault(require("zlib"));
+const tslib_1 = require("tslib");
+const fs_1 = tslib_1.__importDefault(require("fs"));
+const path_1 = tslib_1.__importDefault(require("path"));
+const zlib_1 = tslib_1.__importDefault(require("zlib"));
 // Find the package root (based on the nyc output/ folder)
 const root = (function () {
     let root = process.cwd();