+++ lib.commonjs/index.js
@@ -1,32 +1,7 @@
 "use strict";
-var __createBinding = (this && this.__createBinding) || (Object.create ? (function(o, m, k, k2) {
-    if (k2 === undefined) k2 = k;
-    var desc = Object.getOwnPropertyDescriptor(m, k);
-    if (!desc || ("get" in desc ? !m.__esModule : desc.writable || desc.configurable)) {
-      desc = { enumerable: true, get: function() { return m[k]; } };
-    }
-    Object.defineProperty(o, k2, desc);
-}) : (function(o, m, k, k2) {
-    if (k2 === undefined) k2 = k;
-    o[k2] = m[k];
-}));
-var __setModuleDefault = (this && this.__setModuleDefault) || (Object.create ? (function(o, v) {
-    Object.defineProperty(o, "default", { enumerable: true, value: v });
-}) : function(o, v) {
-    o["default"] = v;
-});
-var __importStar = (this && this.__importStar) || function (mod) {
-    if (mod && mod.__esModule) return mod;
-    var result = {};
-    if (mod != null) for (var k in mod) if (k !== "default" && Object.prototype.hasOwnProperty.call(mod, k)) __createBinding(result, mod, k);
-    __setModuleDefault(result, mod);
-    return result;
-};
-var __exportStar = (this && this.__exportStar) || function(m, exports) {
-    for (var p in m) if (p !== "default" && !Object.prototype.hasOwnProperty.call(exports, p)) __createBinding(exports, m, p);
-};
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.ethers = void 0;
+const tslib_1 = require("tslib");
 /**
  *  The Application Programming Interface (API) is the collection of
  *  functions, classes and types offered by the Ethers library.
@@ -34,7 +9,7 @@ exports.ethers = void 0;
  *  @_section: api:Application Programming Interface  [about-api]
  *  @_navTitle: API
  */
-const ethers = __importStar(require("./ethers.js"));
+const ethers = tslib_1.__importStar(require("./ethers.js"));
 exports.ethers = ethers;
-__exportStar(require("./ethers.js"), exports);
+tslib_1.__exportStar(require("./ethers.js"), exports);
 //# sourceMappingURL=index.js.map