+++ lib.commonjs/providers/provider-ipcsocket.d.ts
@@ -1,4 +1,4 @@
-/// <reference types="node" resolution-mode="require"/>
+/// <reference types="node" />
 import { SocketProvider } from "./provider-socket.js";
 import type { Socket } from "net";
 import type { Networkish } from "./network.js";