+++ lib.commonjs/providers/provider-jsonrpc.js
@@ -279,10 +279,10 @@ class JsonRpcApiProvider extends abstract_provider_js_1.AbstractProvider {
             });
             this.#notReady = { promise, resolve };
         }
-        // This could be relaxed in the future to just check equivalent networks
+        // Make sure any static network is compatbile with the provided netwrok
         const staticNetwork = this._getOption("staticNetwork");
         if (staticNetwork) {
-            (0, index_js_5.assertArgument)(staticNetwork === network, "staticNetwork MUST match network object", "options", options);
+            (0, index_js_5.assertArgument)(network == null || staticNetwork.matches(network), "staticNetwork MUST match network object", "options", options);
             this.#network = staticNetwork;
         }
     }