+++ lib.commonjs/providers/provider-socket.js
@@ -195,23 +195,14 @@ class SocketProvider extends provider_jsonrpc_js_1.JsonRpcApiProvider {
         if ("id" in result) {
             const callback = this.#callbacks.get(result.id);
             if (callback == null) {
-                console.log("Weird... Response for not a thing we sent");
+                this.emit("error", (0, index_js_1.makeError)("received result for unknown id", "UNKNOWN_ERROR", {
+                    reasonCode: "UNKNOWN_ID",
+                    result
+                }));
                 return;
             }
             this.#callbacks.delete(result.id);
             callback.resolve(result);
-            /*
-                        if ("error" in result) {
-                            const { message, code, data } = result.error;
-                            const error = makeError(message || "unkonwn error", "SERVER_ERROR", {
-                                request: `ws:${ JSON.stringify(callback.payload) }`,
-                                info: { code, data }
-                            });
-                            callback.reject(error);
-                        } else {
-                            callback.resolve(result.result);
-                        }
-            */
         }
         else if (result.method === "eth_subscription") {
             const filterId = result.params.subscription;