+++ lib.commonjs/transaction/transaction.js
@@ -142,7 +142,7 @@ function _serializeLegacy(tx, sig) {
     fields.push((0, index_js_3.toBeArray)(sig.s));
     return (0, index_js_3.encodeRlp)(fields);
 }
-function _parseEipSignature(tx, fields, serialize) {
+function _parseEipSignature(tx, fields) {
     let yParity;
     try {
         yParity = handleNumber(fields[0], "yParity");
@@ -181,7 +181,7 @@ function _parseEip1559(data) {
         return tx;
     }
     tx.hash = (0, index_js_2.keccak256)(data);
-    _parseEipSignature(tx, fields.slice(9), _serializeEip1559);
+    _parseEipSignature(tx, fields.slice(9));
     return tx;
 }
 function _serializeEip1559(tx, sig) {
@@ -222,7 +222,7 @@ function _parseEip2930(data) {
         return tx;
     }
     tx.hash = (0, index_js_2.keccak256)(data);
-    _parseEipSignature(tx, fields.slice(8), _serializeEip2930);
+    _parseEipSignature(tx, fields.slice(8));
     return tx;
 }
 function _serializeEip2930(tx, sig) {
@@ -385,7 +385,7 @@ class Transaction {
     get data() { return this.#data; }
     set data(value) { this.#data = (0, index_js_3.hexlify)(value); }
     /**
-     *  The amount of ether to send in this transactions.
+     *  The amount of ether (in wei) to send in this transactions.
      */
     get value() { return this.#value; }
     set value(value) {