+++ lib.commonjs/utils/geturl.js
@@ -1,11 +1,9 @@
 "use strict";
-var __importDefault = (this && this.__importDefault) || function (mod) {
-    return (mod && mod.__esModule) ? mod : { "default": mod };
-};
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.getUrl = void 0;
-const http_1 = __importDefault(require("http"));
-const https_1 = __importDefault(require("https"));
+const tslib_1 = require("tslib");
+const http_1 = tslib_1.__importDefault(require("http"));
+const https_1 = tslib_1.__importDefault(require("https"));
 const zlib_1 = require("zlib");
 const errors_js_1 = require("./errors.js");
 const data_js_1 = require("./data.js");