+++ lib.esm/_admin/update-version.js
@@ -33,6 +33,9 @@ function writeVersion(version) {
         if (log.body.startsWith("admin:")) {
             continue;
         }
+        if (log.body.startsWith("docs:")) {
+            continue;
+        }
         if (log.body.startsWith("tests:")) {
             continue;
         }
@@ -67,6 +70,15 @@ function writeVersion(version) {
                 if (cmp === 0) {
                     return cmp;
                 }
+                // Make sure require comes first; it has the types built-in
+                // so its ok
+                if (a === "require") {
+                    return -1;
+                }
+                if (b === "require") {
+                    return 1;
+                }
+                // Favour types the next-first and default for last
                 if (a === "types" || b === "default") {
                     return -1;
                 }