+++ lib.esm/providers/abstract-provider.js
@@ -79,6 +79,7 @@ async function getSubscription(_event, provider) {
             case "block":
             case "pending":
             case "debug":
+            case "error":
             case "network": {
                 return { type: _event, tag: _event };
             }
@@ -835,6 +836,7 @@ export class AbstractProvider {
     _getSubscriber(sub) {
         switch (sub.type) {
             case "debug":
+            case "error":
             case "network":
                 return new UnmanagedSubscriber(sub.type);
             case "block":