+++ lib.esm/providers/provider-jsonrpc.js
@@ -275,10 +275,10 @@ export class JsonRpcApiProvider extends AbstractProvider {
             });
             this.#notReady = { promise, resolve };
         }
-        // This could be relaxed in the future to just check equivalent networks
+        // Make sure any static network is compatbile with the provided netwrok
         const staticNetwork = this._getOption("staticNetwork");
         if (staticNetwork) {
-            assertArgument(staticNetwork === network, "staticNetwork MUST match network object", "options", options);
+            assertArgument(network == null || staticNetwork.matches(network), "staticNetwork MUST match network object", "options", options);
             this.#network = staticNetwork;
         }
     }